Correção

+++ Circulo/Circulo.java
@@ -8,11 +8,11 @@ public class Circulo {
     }
 
     //método
-    public float calcularArea(){
-        return 3.141516f * raio * raio;
+    public double calcularArea(){
+        return 3.141516 * raio * raio;
     }
-    public float calcularPerimetro(){
-        return 2 * 3.141516f * raio;
+    public double calcularPerimetro(){
+        return 2 * 3.141516 * raio;
     }
     public void imprimir(){
         System.out.printf(" raio: %.2f\n Area: %.2f\n perimetro: %.2f\n ", raio, calcularArea(), calcularPerimetro());
